% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_ec_info}
\alias{all_ec_info}
\title{all element cycle information.}
\format{
a list contains four tables.
\describe{
\item{ec_node}{chemicals}
\item{ec_link}{reactions}
\item{ec_gene}{genes}
\item{ec_path}{reactions labels}
}
}
\description{
all element cycle information.
}
