% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sp_la_zh_name}
\alias{all_sp_la_zh_name}
\title{all species latin names and chinese names}
\format{
a dataframe.
\describe{
\item{latin_name}{latin name}
\item{chinese_name}{chinese name}
}
}
\description{
all species latin names and chinese names.
}
