% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{aor}
\alias{aor}
\alias{aor.data.frame}
\alias{aor.pc_otu}
\alias{plot.AOR}
\title{Calculate Abundance-occupancy_relationship}
\usage{
aor(object, ...)

\method{aor}{data.frame}(otutab, top_r = 0.7, ocup_n = ceiling(0.8 * ncol(otutab)))

\method{aor}{pc_otu}(pc, top_r = 0.7, ocup_n = 0.8, tbl = "otutab")

\method{plot}{AOR}(AOR)
}
\arguments{
\item{object}{object, such as data.frame or pc_otu}

\item{otutab}{otutab}

\item{top_r}{percentage of top relative abundance}

\item{ocup_n}{percentage of top occupied}

\item{AOR}{a AOR object}
}
\value{
AOR
}
\description{
Calculate Abundance-occupancy_relationship

Plot a AOR
}
\examples{
data("otutab")
aor(otutab)->AOR
plot(AOR)
}
