% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{as.dist.b_dist}
\alias{as.dist.b_dist}
\title{Transfer b_dist to dist}
\usage{
\method{as.dist}{b_dist}(aa)
}
\arguments{
\item{aa}{a b_dist object}
}
\value{
dist
}
\description{
Transfer b_dist to dist
}
