% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{b_NTI}
\alias{b_NTI}
\title{Calculate beta_NTI}
\usage{
b_NTI(phylo, otutab, beta.reps = 9, weighted = T, threads = 4)
}
\arguments{
\item{phylo}{a phylo object}

\item{otutab}{otutab}

\item{beta.reps}{how many simulation performed?}

\item{threads}{use how many threads to calculate (default:4)}
}
\value{
a dist: b_NTI
}
\description{
Calculate beta_NTI
}
\examples{
data(otutab)
df2tree(taxonomy)->phylo
b_NTI(phylo,otutab)->bnti_res
data.frame(type=c("Stochastic","Homo_S","Heter_S"),
          number=c(sum(abs(bnti_res)<2),
             sum(bnti_res<(-2)),sum(bnti_res>2)))->com_pro
pcutils::gghuan(com_pro)

}
