% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{b_NTI1}
\alias{b_NTI1}
\title{Calculate beta_NTI}
\usage{
b_NTI1(
  phylo,
  otutab,
  beta.reps = 9,
  weighted = TRUE,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{phylo}{a phylo object}

\item{otutab}{otutab}

\item{beta.reps}{how many simulation performed?}

\item{weighted}{logical}

\item{threads}{use how many threads to calculate (default:4)}

\item{verbose}{verbose}
}
\value{
a dist: b_NTI
}
\description{
Calculate beta_NTI
}
