% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{b_res_3d}
\alias{b_res_3d}
\title{3D plot for b_res}
\usage{
b_res_3d(b_res, Group, metadata = NULL, ...)
}
\arguments{
\item{b_res}{a b_res object}

\item{Group}{group vector for color}

\item{metadata}{metadata contain Group}

\item{...}{add}
}
\value{
plotly list
}
\description{
3D plot for b_res
}
\examples{
data(otutab, package = "pcutils")
b_analyse(otutab, method = "pca", ndim = 3) -> b_res
b_res_3d(b_res, "Group", metadata)
}
