% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{bbtt}
\alias{bbtt}
\title{ggdotchart for diff analysis}
\usage{
bbtt(res, pvalue = "glm.eBH", topN = 20)
}
\arguments{
\item{res}{result of ALDEX or kwtest}

\item{pvalue}{the name of pvaule}

\item{topN}{topN}
}
\value{
ggplot
}
\description{
ggdotchart for diff analysis
}
