% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{before_tree}
\alias{before_tree}
\title{Before df2tree check}
\usage{
before_tree(f_tax)
}
\arguments{
\item{f_tax}{table}
}
\value{
table
}
\description{
Before df2tree check
}
\examples{
wrong_taxdf <- data.frame(
  kingdom = c(rep(c("A", "B"), each = 4), "C", NA),
  "phylum" = c("A", "a", "b", "c", "c", "c", "d", NA, NA, "e")
)
before_tree(wrong_taxdf)

}
