% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{c_means}
\alias{c_means}
\title{C-means cluster}
\usage{
c_means(wtf, k, weight = 1, membership = 0.65)
}
\arguments{
\item{wtf}{standardize data}

\item{k}{cluster number}

\item{weight}{linewidth}

\item{membership}{default:0.65}
}
\value{
ggplot
}
\description{
C-means cluster
}
\examples{
data(otutab,package = "pcutils")
pcutils::hebing(otutab,metadata$Group)->otu_group
group.var= apply(otu_group,1,var)
otu_group.sel = otu_group[group.var >= quantile(group.var, 0.75),]#挑出变化较大的部分，阈值0.75
weight = c(apply(otu_group.sel, 1, var))
wtf = pcutils::trans(otu_group.sel,method = 'standardize',MARGIN = 1)
cm_test_k(wtf)
c_means(wtf,k=3,weight=weight)
}
