% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_net.R
\name{c_net_set}
\alias{c_net_set}
\title{Set basic attributes from totu table}
\usage{
c_net_set(go, totu = NULL, anno_col = NULL, totu2 = NULL, anno_col2 = NULL)
}
\arguments{
\item{go}{a igraph object}

\item{totu}{t(otu table)}

\item{anno_col}{a dataframe using to annotate(with rowname or a name column)}

\item{totu2}{t(otu table) if it exist}

\item{anno_col2}{a dataframe using to annotate(with rowname or a name column)}
}
\value{
a igraph object
}
\description{
Set basic attributes from totu table
}
\examples{
data("otutab")
t(otutab[1:100, ]) -> totu
c_net_cal(totu) -> corr
sample(t(otutab) \%>\% as.data.frame(), 50) -> totu2
colnames(totu2) <- paste0("tt", 1:50)

data("c_net")
co_net <- c_net_set(co_net, t(otutab), taxonomy \%>\% select(Phylum))

c_net_set(co_net2, totu, taxonomy \%>\% select(Phylum), totu2) -> co_net2

}
