% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{check_taxonkit}
\alias{check_taxonkit}
\title{Check taxonkit}
\usage{
check_taxonkit(print = TRUE)
}
\arguments{
\item{print}{print}
}
\value{
taxonkit path
}
\description{
Check taxonkit
}
\seealso{
Other Rtaxonkit: 
\code{\link{download_taxonkit_dataset}()},
\code{\link{install_taxonkit}()},
\code{\link{name_or_id2df}()},
\code{\link{taxonkit_filter}()},
\code{\link{taxonkit_lca}()},
\code{\link{taxonkit_lineage}()},
\code{\link{taxonkit_list}()},
\code{\link{taxonkit_name2taxid}()},
\code{\link{taxonkit_reformat}()}
}
\concept{Rtaxonkit}
