% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{cm_test_k}
\alias{cm_test_k}
\title{Test the proper clusters k for c_means}
\usage{
cm_test_k(wtf, fast = T, choose = NULL)
}
\arguments{
\item{wtf}{standardize data}

\item{fast}{whether do the gap_stat?}

\item{choose}{choose number}
}
\value{
ggplot
}
\description{
Test the proper clusters k for c_means
}
