% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{convert_taxon_name}
\alias{convert_taxon_name}
\title{Convert taxon names between Chinese and Latin}
\usage{
convert_taxon_name(input_names, mode = "latin_to_chinese", fuzzy = FALSE)
}
\arguments{
\item{input_names}{input names}

\item{mode}{conversion mode, "latin_to_chinese" or "chinese_to_latin"}

\item{fuzzy}{whether to use fuzzy matching, default is FALSE}
}
\value{
character vector of converted names
}
\description{
Convert taxon names between Chinese and Latin
}
\examples{
convert_taxon_name(c("Escherichia coli", "Clostridioides difficile"))
}
