% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{cor_net}
\alias{cor_net}
\title{Correlation network, species-interaction network for omics}
\usage{
cor_net()
}
\value{
No value
}
\description{
Correlation network, species-interaction network for omics
}
