% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{df2tree1}
\alias{df2tree1}
\title{From a dataframe to construct a phylo (save nwk)}
\usage{
df2tree1(taxa)
}
\arguments{
\item{taxa}{dataframe}
}
\value{
phylo object
}
\description{
NOTE: this function will transfer all space to \verb{_}
}
\examples{
data(otutab, package = "pcutils")
df2tree1(taxonomy) -> tax_tree
print(tax_tree)
}
