% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{diff_da}
\alias{diff_da}
\title{Difference analysis}
\usage{
diff_da(otutab, group_df, ctrl = NULL, method = "deseq2", log = T)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}

\item{ctrl}{the control group, one level of groups}

\item{method}{one of "deseq2","edger","limma"}

\item{log}{do log transfer for limma?}
}
\value{
a dataframe
}
\description{
Difference analysis
}
\examples{
diff_da(otutab,metadata[,"Group",drop=F])->res
volcano_p(res)
}
