% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{download_taxonkit_dataset}
\alias{download_taxonkit_dataset}
\title{Download taxonkit dataset}
\usage{
download_taxonkit_dataset(make_sure = FALSE, taxdump_tar_gz = NULL)
}
\arguments{
\item{make_sure}{make sure to do this}

\item{taxdump_tar_gz}{your download taxdump_tar_gz file from https://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz}
}
\value{
No value
}
\description{
Download taxonkit dataset
}
