% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{enrich}
\alias{enrich}
\title{GO enrich or KEGG enrich}
\usage{
enrich(geneids, mode = "go")
}
\arguments{
\item{geneids}{geneids}

\item{mode}{go or kegg or wp}
}
\value{
enrich_res
}
\description{
GO enrich or KEGG enrich
}
\examples{
genes=c("ASGR2","BEST1","SIGLEC16","ECRP","C1QC","TCN2","RNASE2","DYSF","C1QB","FAM20A","FCGR1A","CR1","HP","VSIG4","EGR1")
gene2id(genes)->geneid
enrich(geneid$entrezID)->GO
plot.enrich_res(GO$GO_res[1:10,])
enrichplot::cnetplot(GO$GO)
enrichplot::dotplot(GO$GO)
enrichplot::heatplot(GO$GO)
enrichplot::pairwise_termsim(GO$GO)\%>\%enrichplot::emapplot()
enrichplot::upsetplot(GO$GO)
}
