% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{fillNAtax}
\alias{fillNAtax}
\title{Complete a taxonomy table}
\usage{
fillNAtax(
  taxdf,
  taxlevelchar = c("k", "p", "c", "o", "f", "g", "s", "st"),
  na_tax = "Unclassified|uncultured|Ambiguous|Unknown|unknown|metagenome|Unassig",
  ignore.case = TRUE,
  na_repalce = "Unknown"
)
}
\arguments{
\item{taxdf}{taxonomy table}

\item{taxlevelchar}{a vector whose length longer than \code{ncol(taxdf)}, use to be prefix. Default: c("k", "p", "c", "o", "f", "g","s", "st")}

\item{na_tax}{grepl some words and turn to \code{na_repalce}, default: "Unclassified|uncultured|Ambiguous|Unknown|unknown|metagenome|Unassig"}

\item{ignore.case}{ignore.case for \code{na_tax}}

\item{na_repalce}{defalut: Unknown}
}
\value{
a good taxonomy table
}
\description{
Complete a taxonomy table
}
\examples{
taxmat = matrix(sample("onelevel", 7*2, replace = TRUE), nrow = 2, ncol = 7)\%>\%as.data.frame()
colnames(taxmat) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
fillNAtax(taxmat)
}
\references{
\code{MicrobiotaProcess}
}
