% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{gene2id}
\alias{gene2id}
\title{Gene symbol transfer to entrezIDs}
\usage{
gene2id(genes)
}
\arguments{
\item{genes}{gene symbols e.g:ASGR2}
}
\value{
gene entrezIDs dataframe
}
\description{
Gene symbol transfer to entrezIDs
}
\examples{
genes=c("ASGR2","BEST1","SIGLEC16","ECRP","C1QC","TCN2","RNASE2","DYSF","C1QB","FAM20A","FCGR1A","CR1","HP","VSIG4","EGR1")
gene2id(genes)->geneid
}
