% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{get_all_sp_la_zh_name}
\alias{get_all_sp_la_zh_name}
\title{get all species Latin and Chinese name from the CCTCC database}
\usage{
get_all_sp_la_zh_name(
  download_dir = "~/Documents/",
  each_verbose = FALSE,
  max_requests = 50,
  max_id = 30609,
  failure_ids = NULL
)
}
\arguments{
\item{download_dir}{default}

\item{each_verbose}{each_verbose}

\item{max_requests}{default 50}

\item{max_id}{default 30609, try to make sure on the website}

\item{failure_ids}{failure_ids}
}
\value{
No value
}
\description{
get all species Latin and Chinese name from the CCTCC database
}
