% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{get_diff_type}
\alias{get_diff_type}
\title{Get mean and type}
\usage{
get_diff_type(otutab, group_df)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}
}
\description{
Get mean and type
}
