% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{grap_p_test}
\alias{grap_p_test}
\title{Performs graph-based permutation tests}
\usage{
grap_p_test(otutab, metadata, group = "Group", nperm = 999, ...)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{metadata}{metadata}

\item{group}{one group name in columns of metadata}

\item{nperm}{numbers of permutations to perform}

\item{...}{additional}
}
\value{
ggplot
}
\description{
Performs graph-based permutation tests
}
\examples{
\donttest{
if (requireNamespace("phyloseqGraphTest") && requireNamespace("phyloseq")) {
  data(otutab, package = "pcutils")
  grap_p_test(otutab, metadata, "Group")
}
}
}
