% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_plots.R
\name{heat_otu}
\alias{heat_otu}
\title{Heatmap by ggplot}
\usage{
heat_otu(otutab, topN = 20, rowcl = T, colcl = T, r_ann = NULL, c_ann = NULL)
}
\arguments{
\item{otutab}{otutab}

\item{topN}{plot how many top species}

\item{rowcl}{cluster the row?}

\item{colcl}{cluster the column?}

\item{r_ann}{row annotation}

\item{c_ann}{column annotation}
}
\value{
a ggplot
}
\description{
Heatmap by ggplot
}
\examples{
data(otutab)
heat_otu(otutab,topN = 30,r_ann = otutab[1:30,1:2],c_ann = metadata[,c(2,4)])

}
