% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{hebing}
\alias{hebing}
\alias{hebing.data.frame}
\alias{hebing.pc_otu}
\title{Group your data}
\usage{
hebing(object, group, ...)

\method{hebing}{data.frame}(otutab, group, margin = 2, act = "mean")

\method{hebing}{pc_otu}(pc, group, margin = 2, act = "mean", tbl = "otutab")
}
\arguments{
\item{object}{object}

\item{group}{group vector}

\item{...}{additional}

\item{otutab}{dataframe}

\item{margin}{1 for row and 2 for column(default: 2)}

\item{act}{do (default: mean)}
}
\value{
input object
}
\description{
Group your data
}
\examples{
data(otutab)
hebing(otutab,metadata$Group)
}
