% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{install_taxonkit}
\alias{install_taxonkit}
\title{Install taxonkit}
\usage{
install_taxonkit(make_sure = FALSE, taxonkit_tar_gz = NULL)
}
\arguments{
\item{make_sure}{make sure to do this}

\item{taxonkit_tar_gz}{your download taxonkit_tar_gz file from https://github.com/shenwei356/taxonkit/releases/}
}
\value{
No value
}
\description{
Install taxonkit
}
\seealso{
Other Rtaxonkit: 
\code{\link{check_taxonkit}()},
\code{\link{download_taxonkit_dataset}()},
\code{\link{name_or_id2df}()},
\code{\link{taxonkit_filter}()},
\code{\link{taxonkit_lca}()},
\code{\link{taxonkit_lineage}()},
\code{\link{taxonkit_list}()},
\code{\link{taxonkit_name2taxid}()},
\code{\link{taxonkit_reformat}()}
}
\concept{Rtaxonkit}
