% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{kwtest}
\alias{kwtest}
\title{KW test}
\usage{
kwtest(otutab, group_df, method = "kruskal.test")
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}

\item{method}{"kruskal.test", see \code{\link[ggpubr]{compare_means}}}
}
\value{
res
}
\description{
KW test
}
\examples{
\donttest{
data(otutab, package = "pcutils")
kwtest(otutab, metadata["Group"]) -> res
bbtt(res, pvalue = "p.format")
}
}
