% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{life_game}
\alias{life_game}
\title{Life Game Simulation}
\usage{
life_game(file = "LifeGame", time = 100)
}
\arguments{
\item{file}{gif filename}

\item{time}{how many times the life game continue.}
}
\value{
a gif file
}
\description{
Life Game Simulation
}
