% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_phylogenetic.R
\name{makeNewick}
\alias{makeNewick}
\title{From a dataframe to construct a phylo}
\usage{
makeNewick(taxa)
}
\arguments{
\item{taxa}{dataframe}
}
\value{
phylo
}
\description{
From a dataframe to construct a phylo
}
\examples{
data(otutab)
makeNewick(taxonomy)->spe_tree
}
