% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{mat_dist}
\alias{mat_dist}
\title{Calculate distance for otutab}
\usage{
mat_dist(otutab, method = "bray", spe_nwk = NULL)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{method}{Dissimilarity index, partial match to "bray", "euclidean"...see \code{\link[vegan]{vegdist};\link[picante]{unifrac}}}

\item{spe_nwk}{a phylo tree if use unifrac...}
}
\value{
dist
}
\description{
Calculate distance for otutab
}
\examples{
data(otutab,package = "pcutils")
mat_dist(otutab)
}
