% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{multi_bar}
\alias{multi_bar}
\title{Difference analysis}
\usage{
multi_bar(
  otutab,
  group_df,
  mode = 1,
  text_df = NULL,
  text_x = NULL,
  text_angle = -90
)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}

\item{mode}{1~2}

\item{text_df}{text_df}

\item{text_x}{text_x}

\item{text_angle}{text_angle}
}
\value{
a data.frame
}
\description{
Difference analysis
}
\examples{
data(otutab,package = "pcutils")
multi_bar(otutab[1:10,],metadata["Group"])
}
