% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{multi_conf}
\alias{multi_conf}
\title{Difference analysis}
\usage{
multi_conf(otutab, group_df)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}
}
\value{
ggplot
}
\description{
Difference analysis
}
\examples{
data(otutab, package = "pcutils")
multi_conf(otutab[1:10, 1:12], metadata["Group"])
}
