% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstics.R
\name{multireg}
\alias{multireg}
\title{Multiple regression/ variance decomposition analysis}
\usage{
multireg(formula, data, TopN = 3)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe}

\item{TopN}{give top variable importance}
}
\description{
Multiple regression/ variance decomposition analysis
}
\examples{
data(otutab)
cbind(a_diversity(otutab,method="shannon"),metadata[,c(2,12:15)])->a_res
multireg(formula = Shannon~Group*.,data = a_res)
}
