% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{myRDA}
\alias{myRDA}
\alias{myCCA}
\alias{myCAP}
\title{RDA}
\usage{
myRDA(otutab, env, choose_var = F, norm = T, scale = F)

myCCA(otutab, env, choose_var = F, norm = T, scale = F)

myCAP(otutab, env, choose_var = F, norm = T, scale = F, dist = "bray")
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{env}{environmental factors}

\item{choose_var}{should choose variables? use forward step}

\item{norm}{should normalize?(default:T)}

\item{scale}{should scale species?(default:F)}

\item{dist}{dist method for CAP, see \code{\link[vegan]{vegdist}}}
}
\value{
rda/cca
}
\description{
RDA
}
\examples{
data(otutab)
env=metadata[,12:16]
#RDA
myRDA(otutab,env)->phy.rda
RDA_plot(phy.rda,metadata$Group)
}
\seealso{
\code{\link[vegan]{vegdist};\link[picante]{unifrac}}
}
