% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_net.R
\name{net_par}
\alias{net_par}
\title{Calculate all indexs of a network}
\usage{
net_par(p, mode = c("v", "e", "n", "all"))
}
\arguments{
\item{p}{a igraph object}

\item{mode}{calculate what? c("v", "e", "n", "all")}
}
\value{
a 3-elements list
\item{n_index}{indexs of the whole network}
\item{v_index}{indexs of each vertex}
\item{e_index}{indexs of each edge}
}
\description{
Calculate all indexs of a network
}
\examples{
make_graph("Walther") \%>\% net_par()
}
