% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{nti_rc}
\alias{nti_rc}
\title{Calculate b_NTI and RC_bray for each group}
\usage{
nti_rc(otutab, phylo, group_df, threads = 4, file = NULL, rep = 20)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{phylo}{a phylo object}

\item{group_df}{a dataframe with rowname and one group column}

\item{threads}{default:4}

\item{file}{filename to save}

\item{rep}{repeat numbers: suggest 999}
}
\value{
a b_dist object, dis is MSTij
}
\description{
Calculate b_NTI and RC_bray for each group
}
\examples{
data(otutab)
df2tree(taxonomy)->phylo
nti_rc(otutab,phylo,metadata[,"Group",drop=F])->nti_res
nti_res$type=factor(nti_res$type,levels = c("Homo_S","Heter_S","Homo_D","D_limit","Undominated"))
table(nti_res$type,nti_res$variable)\%>\%reshape2::melt()->com_p
colnames(com_p)=c("type","variable","n")
ggplot(com_p,aes(x=variable,y=n))+geom_bar(stat = "identity",aes(fill=type),position = "fill")

}
\references{
1. Ning, D., Deng, Y., Tiedje, J. M. & Zhou, J. A general framework for quantitatively assessing ecological stochasticity. Proceedings of the National Academy of Sciences 116, 16892–16898 (2019).
}
