% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_otu.R
\name{pc_otu}
\alias{pc_otu}
\title{Create a pc_otu class object}
\usage{
pc_otu(otutab = data.frame(), metadata = data.frame(), taxonomy = NULL, ...)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{metadata}{a metadata data.frame, samples are rows}

\item{taxonomy}{a taxomomy data.frame, look out the rowname of taxonomy and otutab should matched!}
}
\value{
pc_otu
}
\description{
Create a pc_otu class object
}
\examples{
data("otutab")
otu1<-pc_otu(otutab,metadata)
otu1
}
