% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_tax1}
\alias{pc_tax1}
\title{test data (pc_otu class) for pc_tax package.}
\format{
a pc_otu contains an otutab, metadata and a taxonomy table.
\describe{
\item{tbls}{contians otutable rawdata}
\item{metas}{contians metadata}
\item{otus}{contians taxomomy table}
}
}
\description{
an otutab, metadata and a taxonomy table.
}
