% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctax-package.R
\docType{package}
\name{pctax-package}
\alias{pctax}
\alias{pctax-package}
\title{pctax: Professional Comprehensive Microbiome Data Analysis}
\description{
Provides a comprehensive suite of tools for analyzing microbiome data. It includes functionalities for alpha diversity analysis, beta diversity analysis, differential analysis, community assembly, visualization of phylogenetic tree and functional enrichment analysis... With a progressive approach, the package offers a range of analysis methods to explore and understand the complex microbial communities. It is designed to support researchers and practitioners in conducting in-depth and professional microbiome data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/pctax}
  \item Report bugs at \url{https://github.com/Asa12138/pctax/issues}
}

}
\author{
\strong{Maintainer}: Peng Chen \email{pengchen2001@zju.edu.cn}

}
\keyword{internal}
