% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{permanova2}
\alias{permanova2}
\title{Permanova between a otutab and a variable (added \code{two})}
\usage{
permanova2(
  otutab,
  envs,
  norm = TRUE,
  each = TRUE,
  method = "adonis",
  dist = "bray",
  two = FALSE,
  nperm = 999,
  ...
)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{envs}{factors need to test}

\item{norm}{should normalize?(default:TRUE)}

\item{each}{test factor one by one, rather than whole}

\item{method}{adonis/mrpp/anosim/mantel}

\item{dist}{if use pcoa or nmds, your can choose a dist method (default: bray)}

\item{two}{two by two adonis test}

\item{nperm}{numbers of permutations to perform}

\item{...}{additional}
}
\description{
Permanova between a otutab and a variable (added \code{two})
}
