% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot.b_res}
\alias{plot.b_res}
\title{Plot a b_res}
\usage{
\method{plot}{b_res}(
  x,
  Group,
  metadata = NULL,
  Group2 = NULL,
  mode = 1,
  bi = FALSE,
  Topn = 10,
  rate = 1,
  margin = FALSE,
  margin_label = TRUE,
  permanova_res = NULL,
  text_param = list(),
  box_margin = TRUE,
  box_param = list(),
  pal = NULL,
  sample_label = TRUE,
  stat_ellipse = TRUE,
  ellipse_level = 0.95,
  add_centroid_label = TRUE,
  coord_fix = FALSE,
  bi_text_size = 3,
  ...
)
}
\arguments{
\item{x}{a b_res object}

\item{Group}{group vector for color}

\item{metadata}{metadata contain Group}

\item{Group2}{mapping point shape}

\item{mode}{plot mode:1~3}

\item{bi}{plot variables segments?}

\item{Topn}{how many variables to show?}

\item{rate}{segments length rate}

\item{margin}{plot the margin boxplot?}

\item{margin_label}{margin_label, TRUE}

\item{permanova_res}{permanova result}

\item{text_param}{text_param for \code{\link[ggplot2]{annotate}}}

\item{box_margin}{margin plot box or density?}

\item{box_param}{box_param for \code{\link[pcutils]{group_box}}}

\item{pal}{colors for group}

\item{sample_label}{plot the labels of samples?}

\item{stat_ellipse}{plot the stat_ellipse?}

\item{ellipse_level}{the level of stat_ellipse, defalut 0.95}

\item{add_centroid_label}{add the centroid label in mode 3?}

\item{coord_fix}{fix the coordinates y/x ratio}

\item{bi_text_size}{biplot text size}

\item{...}{add}
}
\value{
a ggplot
}
\description{
Plot a b_res
}
\seealso{
\code{\link{b_analyse}}
}
