% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{plot.enrich_res}
\alias{plot.enrich_res}
\title{Plot enrich_res}
\usage{
\method{plot}{enrich_res}(x, mode = 1, str_width = 50, ...)
}
\arguments{
\item{x}{enrich_res object}

\item{mode}{mode}

\item{str_width}{default: 50}

\item{...}{add}
}
\value{
ggplot
}
\description{
Plot enrich_res
}
