% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{plot.ncm_res}
\alias{plot.ncm_res}
\title{Plot ncm_res}
\usage{
\method{plot}{ncm_res}(ncm_res)
}
\arguments{
\item{res}{a ncm_res object}
}
\value{
ggplot
}
\description{
Plot ncm_res
}
\seealso{
\code{\link{ncm}}
}
