% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot.pro_res}
\alias{plot.pro_res}
\title{Plot pro_res}
\usage{
\method{plot}{pro_res}(x, group, metadata = NULL, pal = NULL, ...)
}
\arguments{
\item{x}{pro_res}

\item{group}{group}

\item{metadata}{metadata}

\item{pal}{pal}

\item{...}{add}
}
\value{
a ggplot
}
\description{
Plot pro_res
}
