% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot_b_like}
\alias{plot_b_like}
\title{base plot for pca/rda}
\usage{
plot_b_like(
  plotdat,
  mode = 1,
  pal = NULL,
  sample_label = T,
  groupname = "level",
  groupname2 = "level2"
)
}
\description{
base plot for pca/rda
}
