% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_software.R
\name{pre_assembly_stats}
\alias{pre_assembly_stats}
\title{Prepare the result from assembly_stats (.json file)}
\usage{
pre_assembly_stats(jsonfiles)
}
\arguments{
\item{jsonfiles}{the directory contains .json file}
}
\value{
data.frame
}
\description{
Prepare the result from assembly_stats (.json file)
}
