% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_software.R
\name{pre_fastp}
\alias{pre_fastp}
\title{Prepare the result from fastp (.json file)}
\usage{
pre_fastp(jsonfiles, prefix = c("Raw", "Clean"))
}
\arguments{
\item{jsonfiles}{the directory contains .json file}

\item{prefix}{default c("Raw","Clean"), for the before filtering and after filtering.}
}
\value{
data.frame
}
\description{
Prepare the result from fastp (.json file)
}
