% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctax_additional.R
\name{pre_GEO}
\alias{pre_GEO}
\title{prepare the GEO data}
\usage{
pre_GEO(my_id, GEO_dir = "GEO_data", file = NULL)
}
\arguments{
\item{my_id}{GEOid}

\item{GEO_dir}{GEO download dir}

\item{file}{the downloaded file}
}
\value{
list(meta = meta, GSE_expr = GSE_expr)
}
\description{
prepare the GEO data
}
