% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_software.R
\name{print.virus_res}
\alias{print.virus_res}
\title{Print method for virus_res objects}
\usage{
\method{print}{virus_res}(x, ...)
}
\arguments{
\item{x}{An object of class virus_res}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for virus_res objects
}
