% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{procrustes_analyse}
\alias{procrustes_analyse}
\title{Procrustes Rotation of Two Configurations and PROTEST}
\usage{
procrustes_analyse(b_res1, b_res2)
}
\arguments{
\item{b_res1}{Target matrix}

\item{b_res2}{Matrix to be rotated}
}
\value{
pro_res
}
\description{
Procrustes Rotation of Two Configurations and PROTEST
}
\examples{
data(otutab, package = "pcutils")
b_analyse(otutab, method = "pca") -> b_res1
b_analyse(otutab * abs(rnorm(10)), method = "pca") -> b_res2
pro_res <- procrustes_analyse(b_res1, b_res2)
plot(pro_res, "Group", metadata)
}
