% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rare_curves}
\alias{rare_curves}
\title{Plot a rare curve}
\usage{
rare_curves(otutab, step = 40000, method = "richness", mode = 2, threads = 4)
}
\arguments{
\item{otutab}{otutab}

\item{step}{default 20000}

\item{method}{one of "richness","chao1","ace","gc","shannon","simpson","pd","pielou"}

\item{mode}{1 for little table, 2 for big}

\item{threads}{use how many threads to calculate (default:4)}
}
\description{
Plot a rare curve
}
\examples{
data(otutab)
rare_curves(otutab,mode=1)
}
