% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefaction}
\alias{rarefaction}
\title{Rarefy a otutab}
\usage{
rarefaction(otutab, sample = NULL)
}
\arguments{
\item{otutab}{otutab}

\item{sample}{number}
}
\value{
a rarefied otutab
}
\description{
Rarefy a otutab
}
\examples{
data(otutab,package = "pcutils")
rarefaction(otutab)
}
