% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{RDA_plot}
\alias{RDA_plot}
\title{Plot RDA res}
\usage{
RDA_plot(
  phy.rda,
  Group,
  mode = 1,
  tri = F,
  Topn = 10,
  rate = 1,
  pal = NULL,
  sample_label = T,
  groupname = "level"
)
}
\arguments{
\item{phy.rda}{rda/cca object}

\item{Group}{group vector to color}

\item{mode}{which mode choosen to plot}

\item{tri}{plot species arrows?}

\item{Topn}{how many variables to show?}

\item{rate}{segments length rate}

\item{pal}{colors for group}

\item{sample_label}{show the samples label?}
}
\value{
ggplot
}
\description{
Plot RDA res
}
\seealso{
\code{\link{myRDA}}
}
