% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_net.R
\name{robustness_test}
\alias{robustness_test}
\alias{plot.robustness}
\title{Robustness_test for a network}
\usage{
robustness_test(go, partial = 0.7, step = 1, reps = 9, threads = 4)

\method{plot}{robustness}(robust_res, indexs = c("n_conn", "aplength", "degree"))
}
\arguments{
\item{go}{a igraph object}

\item{partial}{how much percent vertexes be removed}

\item{robust_res}{robustness_test() result (robustness class)}

\item{indexs}{indexs selected to show}
}
\value{
dataframe(robustness class)

a ggplot
}
\description{
Robustness_test for a network

Plot robustness
}
\examples{
data("c_net")
robustness_test(co_net,threads=1)->robust_res
class(robust_res)
plot(robust_res)
}
