% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_phylogenetic.R
\name{sangji_plot}
\alias{sangji_plot}
\alias{sunburst}
\title{Plot a sankey}
\usage{
sangji_plot(
  tree,
  top_N = 5,
  notshow = c(),
  intermediate = F,
  width = 3000,
  height = 500,
  ...
)

sunburst(tree)
}
\arguments{
\item{tree}{ann_tree result}

\item{top_N}{each level has top_N}

\item{notshow}{some words you don't want to show}
}
\value{
sunburst
}
\description{
Plot a sankey

Plot a sunburst
}
\examples{
data(otutab)
ann_tree(taxonomy[,c(1,5,6,7)],otutab)->tree
sangji_plot(tree)
sunburst(tree)
}
