% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_plots.R
\name{stackplot}
\alias{stackplot}
\alias{areaplot}
\alias{tax_circlize}
\title{Plot a stack plot,}
\usage{
stackplot(
  otutab,
  metadata = NULL,
  topN = 8,
  groupID = "Group",
  shunxu = F,
  relative = T,
  style = "group",
  sorted = "abundance",
  flow = F,
  others = T,
  pmode = "stack"
)

areaplot(
  otutab,
  metadata,
  topN = 8,
  groupID = "Date",
  relative = T,
  sorted = "abundance",
  others = T
)

tax_circlize(
  otutab,
  metadata,
  topN = 5,
  groupID = "Group",
  others = T,
  rowcol = c(),
  colcol = c()
)
}
\arguments{
\item{otutab}{otutab}

\item{metadata}{metadata}

\item{topN}{plot how many top species}

\item{groupID}{one group name of columns of metadata}

\item{shunxu}{should order the samples by the top1 abundance}

\item{relative}{transfer to relative or absolute}

\item{style}{"group" or "sample"}

\item{sorted}{should legend be sorted by "abunance"}

\item{flow}{should plot a flow plot?}

\item{others}{should plot others?}

\item{pmode}{fill/stack/dodge}

\item{rowcol}{colors codes}

\item{colcol}{colors codes}
}
\value{
plot
}
\description{
Plot a stack plot,

Areaplot

Circle plot
}
\examples{
data(otutab)
stackplot(otutab,metadata,groupID="Group")
stackplot(otutab,metadata,groupID="Group",shunxu=T,flow=T,relative=F)

hclust(dist(t(otutab)))\%>\%ape::as.phylo()\%>\%as_tibble()->s_tree
full_join(s_tree,metadata,by=c("label"="Id"))->s_tree
library(ggtree)
ggtree(tidytree::as.treedata(s_tree))+geom_tippoint(aes(col=Group,shape=Group),size=2)->p1
stackplot(hebing(otutab,taxonomy$Phylum,1),metadata,groupID ='Id',topN = 10,others = T,flow = T)+
  coord_flip()+ scale_y_continuous(expand=c(0,0)) +xlab("")->pp2
pp2\%>\%aplot::insert_left(p1, width=.3)

areaplot(otutab,metadata,topN = 10,others = F)
tax_circlize(otutab,metadata,groupID ='Site',topN = 10,others = F)
}
