% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{stamp_plot}
\alias{stamp_plot}
\title{Stamp style plot}
\usage{
stamp_plot(otutab, group_df, set_order = NULL, pal = NULL)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}

\item{set_order}{set order of factor levels}

\item{pal}{palette}
}
\value{
ggplot
}
\description{
Stamp style plot
}
\examples{
data(otutab, package = "pcutils")
if (requireNamespace("ggfun")) stamp_plot(otutab[1:10, 1:12], metadata["Group"])
}
