% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{suijisenlin}
\alias{suijisenlin}
\title{RandomForest}
\usage{
suijisenlin(otutab, group_df, topN = 10)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}

\item{topN}{default: 10}
}
\value{
diff
}
\description{
RandomForest
}
\examples{
data(otutab,package = "pcutils")
suijisenlin(otutab,metadata["Group"])->rf_res
}
