% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_otu.R
\name{summary.pc_otu}
\alias{summary.pc_otu}
\title{Summary pc_otu}
\usage{
\method{summary}{pc_otu}(object, ...)
}
\arguments{
\item{object}{pc_otu}

\item{...}{add}
}
\value{
No value
}
\description{
Summary pc_otu
}
\examples{
data("pc_tax1")
summary(pc_tax1)

}
