% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{sunburst}
\alias{sunburst}
\title{Plot a sunburst}
\usage{
sunburst(tree)
}
\arguments{
\item{tree}{result from \code{\link{ann_tree}}}
}
\value{
sunburst
}
\description{
Plot a sunburst
}
\examples{
\donttest{
data(otutab, package = "pcutils")
ann_tree(taxonomy[, c(1, 5, 6, 7)], otutab) -> tree
sunburst(tree)
}
}
\seealso{
\code{\link[=sangji_plot]{sangji_plot()}}
}
