% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{tax_lca}
\alias{tax_lca}
\title{Calculate the lowest common ancestor (LCA) of a set of taxa}
\usage{
tax_lca(df)
}
\arguments{
\item{df}{a data frame with taxonomic information, with columns representing taxonomic levels}
}
\value{
character
}
\description{
Calculate the lowest common ancestor (LCA) of a set of taxa
}
\examples{
df <- data.frame(
  A = c("a", "a", "a", "a"),
  B = c("x", "x", "y", "y"),
  C = c("1", "1", "2", "3"),
  stringsAsFactors = FALSE
)
tax_lca(df)
}
