% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_plots.R
\name{tax_pie}
\alias{tax_pie}
\title{Pie plot}
\usage{
tax_pie(otutab, n = 6)
}
\arguments{
\item{otutab}{otutab}

\item{n}{topn}
}
\value{
ggplot
}
\description{
Pie plot
}
\examples{
tax_pie(otutab,n = 7)
}
