% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{time_by_cm}
\alias{time_by_cm}
\title{Time series analysis}
\usage{
time_by_cm(otu_time, n_cluster = 6, mem_thr = 0.6, min.std = 0)
}
\arguments{
\item{otu_time}{otutab hebing by a time variable}

\item{n_cluster}{number of clusters}

\item{mem_thr}{membership threshold}

\item{min.std}{}
}
\description{
Time series analysis
}
\examples{
otu_time=hebing(otutab,metadata$Group)
time_by_cm(otu_time,n_cluster=4)
}
