% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\alias{venn.list}
\alias{venn.data.frame}
\alias{venn.pc_otu}
\title{plot a general venn (upset, flower)}
\usage{
venn(object, ...)

\method{venn}{list}(aa, mode = "venn", ...)

\method{venn}{data.frame}(otutab, mode = "venn")

\method{venn}{pc_otu}(pc, mode = "venn", tbl = "otutab")
}
\arguments{
\item{object}{list/data.frame/pc_otu}

\item{...}{additional}

\item{mode}{"venn","venn2","upset","flower"}
}
\value{
a plot
}
\description{
plot a general venn (upset, flower)
}
\examples{
aa=list(a=1:3,b=3:7,c=2:4)
venn(aa)
}
