% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_cal}
\alias{venn_cal}
\title{Calculate dataframe rownames(>0) in each columns}
\usage{
venn_cal(otu_time)
}
\arguments{
\item{otu_time}{a dataframe}
}
\value{
a list
}
\description{
Calculate dataframe rownames(>0) in each columns
}
\examples{
data.frame(row.names = letters[1:10],col1=sample(0:1,10,T),col2=sample(0:1,10,T))->aa
venn_cal(aa)
}
