% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{volcano_p}
\alias{volcano_p}
\title{Volcano plot for difference analysis}
\usage{
volcano_p(
  res,
  logfc = 1,
  adjp = 0.05,
  text = TRUE,
  repel = TRUE,
  mode = 1,
  number = FALSE
)
}
\arguments{
\item{res}{result of \code{diff_da} which have colnames: tax, log2FoldChange, padj, compare, sig}

\item{logfc}{log_fold_change threshold}

\item{adjp}{adjust_p_value threshold}

\item{text}{text, TRUE}

\item{repel}{repel, TRUE}

\item{mode}{1:normal; 2:multi_contrast}

\item{number}{show the tax number}
}
\value{
ggplot
}
\description{
Volcano plot for difference analysis
}
\seealso{
\code{\link{diff_da}}
}
