% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_net.R
\name{zp_plot}
\alias{zp_plot}
\title{Zi-Pi plot of vertexes}
\usage{
zp_plot(indexs)
}
\arguments{
\item{indexs}{indexs in modu_dect() result}
}
\value{
a ggplot object
}
\description{
Zi-Pi plot of vertexes
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
zp_plot(co_net_modu$indexs)
}
