% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_diversity.R
\name{a_diversity}
\alias{a_diversity}
\alias{a_diversity.data.frame}
\alias{a_diversity.pc_otu}
\alias{a_diversity.numeric}
\title{Calculate a_diversity of otutab}
\usage{
a_diversity(object, ...)

\method{a_diversity}{data.frame}(otutab, method = c("simpson", "shannon"), tree = NULL, digits = 4)

\method{a_diversity}{pc_otu}(pc, method = "all", tbl = "otutab")

\method{a_diversity}{numeric}(x, method = c("simpson", "shannon"), tree = NULL)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{method}{one of "all","richness","chao1","ace","gc","shannon","simpson","pd","pielou"}

\item{tree}{a iphylo object match the rownames of otutab}

\item{digits}{maintance how many digits}
}
\value{
a a_res object
}
\description{
Calculate a_diversity of otutab
}
\examples{
data(otutab)
a_diversity(otutab)->a_res
plot(a_res,metadata,"Group")
plot(a_res,metadata,"env1")
}
