% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_otu.R
\name{add_tax}
\alias{add_tax}
\title{Add taxonomy for a pc_otu object}
\usage{
add_tax(pc, taxonomy)
}
\arguments{
\item{pc}{a pc_otu object}

\item{taxonomy}{a taxomomy data.frame, look out the rownames of taxonomy and otutab should matched!}
}
\value{
pc_otu
}
\description{
Add taxonomy for a pc_otu object
}
\examples{
data(otutab, package = "pcutils")
pc_tax1 <- pc_otu(otutab, metadata)
pc_tax1 <- add_tax(pc_tax1, taxonomy)
}
