% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{ALDEX}
\alias{ALDEX}
\title{ALDEX}
\usage{
ALDEX(otutab, group_df)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}
}
\value{
diff
}
\description{
ALDEX
}
\examples{
\donttest{
data(otutab, package = "pcutils")
ALDEX(otutab, metadata["Group"]) -> res
res \%>\%
    dplyr::top_n(9, -glm.eBH) \%>\%
    .[, "tax"] -> sig
data.frame(t(otutab[sig, ])) \%>\% pcutils::group_box(., "Group", metadata)
}
}
\references{
\url{https://cloud.tencent.com/developer/article/1621879}
}
