% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{ALDEX}
\alias{ALDEX}
\title{ALDEX}
\usage{
ALDEX(otutab, group_df)
}
\arguments{
\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}
}
\value{
diff
}
\description{
ALDEX
}
\examples{
ALDEX(otutab,metadata[,"Group",drop=F])->res
res\%>\%top_n(9,-glm.eBH)\%>\%pull(tax)->sig
data.frame(t(otutab[sig,]))\%>\%group_box(.,group = metadata$Group,p_value = T)
}
\references{
https://cloud.tencent.com/developer/article/1621879
}
