% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{as.dist.b_dist}
\alias{as.dist.b_dist}
\title{Transfer b_dist to dist}
\usage{
\method{as.dist}{b_dist}(m, diag = FALSE, upper = FALSE)
}
\arguments{
\item{m}{a b_dist object}

\item{diag}{logical value indicating whether the diagonal of the distance matrix should be printed by \code{print.dist}.}

\item{upper}{logical value indicating whether the upper triangle of the distance matrix should be printed by \code{print.dist}.}
}
\value{
dist
}
\description{
Transfer b_dist to dist
}
