% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{b_analyse}
\alias{b_analyse}
\alias{b_analyse.data.frame}
\alias{b_analyse.pc_otu}
\title{Beta_diversity Ordination: dimensionality reduction}
\usage{
b_analyse(object, ...)

\method{b_analyse}{data.frame}(
  otutab,
  norm = T,
  method = c("pca", "nmds"),
  group = NULL,
  dist = "euclidean",
  ndim = 2
)

\method{b_analyse}{pc_otu}(
  pc,
  tbl = "otutab",
  norm = T,
  method = c("pca", "ca"),
  group = NULL,
  dist = "euclidean"
)
}
\arguments{
\item{object}{object}

\item{...}{additional}

\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{norm}{should normalized or not? (hellinger)}

\item{method}{one of "pca","pcoa","ca","dca","nmds","plsda","tsne","umap","lda","all"}

\item{group}{if needed, give a group vector}

\item{dist}{if use pcoa, your can choose a dist method(default: euclidean)}

\item{ndim}{how many dimension be kept?(default:2) 3 for b_res_3d()}
}
\value{
b_res object
}
\description{
Beta_diversity Ordination: dimensionality reduction
}
\examples{
data(otutab)
b_analyse(otutab,method = "pca")->b_res
plot(b_res,metadata$Group)
#b_analyse(otutab,method = "all",group=metadata$Group)->b_res
#plot(b_res,metadata$Group,mode=3)
}
\references{
\link{https://www.jianshu.com/p/9694c0b6302d}
\link{https://mixomicsteam.github.io/Bookdown/plsda.html}
\link{https://zhuanlan.zhihu.com/p/25501130}
}
