% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{b_dist}
\alias{b_dist}
\title{Calulate distance for otutab}
\usage{
b_dist(otutab, method = "bray", spe_nwk = NULL)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{method}{Dissimilarity index, partial match to "bray", "euclidean"...}

\item{spe_nwk}{a phylo tree}
}
\value{
dist
}
\description{
Calulate distance for otutab
}
\examples{
data(otutab)
b_dist(otutab)
makeNewick(taxonomy)->spe_nwk
b_dist(otutab)
}
\seealso{
\code{\link[vegan]{vegdist};\link[picante]{unifrac}}
}
