% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{b_res_3d}
\alias{b_res_3d}
\title{3D plot for b_res}
\usage{
b_res_3d(b_res, group)
}
\arguments{
\item{b_res}{a b_res object}

\item{group}{group vector for color}
}
\value{
plotly list
}
\description{
3D plot for b_res
}
\examples{
data(otutab)
b_analyse(otutab,method = "pca",ndim=3)->b_res
b_res_3d(b_res,metadata$Group)
}
