% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{bbtt}
\alias{bbtt}
\title{ggdotchart for diff}
\usage{
bbtt(res, pvalue = "glm.eBH", topN = 20)
}
\arguments{
\item{res}{result of ALDEX or kwtest}

\item{pvalue}{the name of pvaule}
}
\value{
ggplot
}
\description{
ggdotchart for diff
}
\examples{
ALDEX(otutab,metadata[,"Group",drop=F])->res
bbtt(res)
}
