% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{c_means}
\alias{c_means}
\title{C-means cluster}
\usage{
c_means(otutab, wtf, k, weight = 1)
}
\arguments{
\item{otutab}{original data}

\item{wtf}{standardize data}

\item{k}{cluster number}

\item{weight}{linewidth}
}
\value{
ggplot
}
\description{
C-means cluster
}
\examples{
hebing(otutab,group)->date_gen_cpm
season.var = apply(date_gen_cpm,1,var)
df.season.sel = date_gen_cpm[season.var >= quantile(season.var, 0.75),]#挑出变化较大的部分，阈值0.75
weight = c(apply(df.season.sel, 1, var))
wtf = decostand(df.season.sel,method = 'standardize',MARGIN = 1)
cm_test_k(wtf)
c_means(otutab,wtf,k=3,weight)
}
