% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{plot.time_cm}
\alias{plot.time_cm}
\title{Plot time_cm}
\usage{
\method{plot}{time_cm}(x, mem_thr = 0.6, ...)
}
\arguments{
\item{x}{time_cm}

\item{mem_thr}{membership threshold}

\item{...}{add}
}
\value{
ggplot
}
\description{
Plot time_cm
}
