% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{c_means}
\alias{c_means}
\title{C-means cluster}
\usage{
c_means(otu_group, k_num, filter_var)
}
\arguments{
\item{otu_group}{standardize data}

\item{k_num}{cluster number}

\item{filter_var}{filter the highest var}
}
\value{
ggplot
}
\description{
C-means cluster
}
\examples{
data(otutab,package = "pcutils")
pcutils::hebing(otutab,metadata$Group)->otu_group
cm_test_k(otu_group,filter_var=0.7)
cm_res=c_means(otu_group,k_num=3,filter_var=0.7)
}
