% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{cm_test_k}
\alias{cm_test_k}
\title{Test the proper clusters k for c_means}
\usage{
cm_test_k(wtf, fast = T)
}
\arguments{
\item{fast}{whether do the gap_stat?}

\item{otutab}{otutab}

\item{group_df}{a dataframe with rowname same to dist and one group column}
}
\value{
ggplot
}
\description{
Test the proper clusters k for c_means
}
