% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{cm_test_k}
\alias{cm_test_k}
\title{Test the proper clusters k for c_means}
\usage{
cm_test_k(otu_group, filter_var, fast = T)
}
\arguments{
\item{otu_group}{grouped otutab}

\item{filter_var}{filter the highest var}

\item{fast}{whether do the gap_stat?}
}
\value{
ggplot
}
\description{
Test the proper clusters k for c_means
}
