% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{df2tree}
\alias{df2tree}
\title{From a dataframe to construct a phylo}
\usage{
df2tree(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
phylo object
}
\description{
From a dataframe to construct a phylo
}
\examples{
data(otutab,package = "pcutils")
df2tree(taxonomy)->tax_tree
print(tax_tree)
#check all nodes matched!
picante::match.phylo.comm(tax_tree,t(otutab))->nn
nrow(nn$comm)==nrow(t(otutab))
}
