% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{download_taxonkit_dataset}
\alias{download_taxonkit_dataset}
\title{Download taxonkit dataset}
\usage{
download_taxonkit_dataset(taxdump_tar_gz = NULL)
}
\arguments{
\item{taxdump_tar_gz}{your download taxdump_tar_gz file from ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz}
}
\description{
Download taxonkit dataset
}
\examples{
\dontrun{
download_taxonkit_dataset()
}
}
