% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{envfitt}
\alias{envfitt}
\title{Envfit test for RDA result}
\usage{
envfitt(phy.rda, env, ...)
}
\arguments{
\item{phy.rda}{a rda result}

\item{env}{environmental factors}

\item{...}{add}
}
\value{
g_test object
}
\description{
Envfit test for RDA result
}
\examples{
data(otutab,package = "pcutils")
env=env=metadata[,6:10]
#RDA
myRDA(otutab,env)->phy.rda
envfitt(phy.rda,env)
}
\seealso{
\code{\link[vegan]{envfit}}
}
