% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_phylogenetic.R
\name{fillNAtax}
\alias{fillNAtax}
\title{Complete a taxonomy table}
\usage{
fillNAtax(taxdf, type = "species")
}
\arguments{
\item{taxdf}{taxonomy table}

\item{type}{species}
}
\value{
a good taxonomy table
}
\description{
Complete a taxonomy table
}
\examples{
taxmat = matrix(sample("onelevel", 7*2, replace = TRUE), nrow = 2, ncol = 7)\%>\%as.data.frame()
colnames(taxmat) <- c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
fillNAtax(taxmat)
}
\references{
MicrobiotaProcess
}
