% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctax_additional.R
\name{gene2id}
\alias{gene2id}
\title{Gene symbolid transfer to entrezIDs (human gene)}
\usage{
gene2id(genes)
}
\arguments{
\item{genes}{gene symbols e.g:ASGR2}
}
\value{
gene entrezIDs dataframe
}
\description{
Gene symbolid transfer to entrezIDs (human gene)
}
\examples{
if (requireNamespace("AnnotationDbi") && requireNamespace("org.Hs.eg.db")) {
  genes <- c(
    "ASGR2", "BEST1", "SIGLEC16", "ECRP", "C1QC", "TCN2", "RNASE2",
    "DYSF", "C1QB", "FAM20A", "FCGR1A", "CR1", "HP", "VSIG4", "EGR1"
  )
  gene2id(genes) -> geneid
}
}
