% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{geo_sim}
\alias{geo_sim}
\title{Lm for sample similarity and geographical distance}
\usage{
geo_sim(otutab, geo, ...)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{geo}{a two-columns dataframe, first is latitude, second is longitude}

\item{...}{additional}
}
\value{
a ggplot
}
\description{
Lm for sample similarity and geographical distance
}
\examples{
data(otutab)
data.frame(row.names = rownames(metadata),lat=runif(18,30,35),long=runif(18,40,45))->geo
geo_sim(otutab,geo)
}
