% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{install_taxonkit}
\alias{install_taxonkit}
\title{Install taxonkit}
\usage{
install_taxonkit(taxonkit_tar_gz = NULL)
}
\arguments{
\item{taxonkit_tar_gz}{your download taxonkit_tar_gz file from https://github.com/shenwei356/taxonkit/releases/}
}
\description{
Install taxonkit
}
\examples{
\dontrun{
install_taxonkit()
}
}
