% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analyse.R
\name{kwtest}
\alias{kwtest}
\title{Title}
\usage{
kwtest(otutab, group_df)
}
\arguments{
\item{group_df}{}
}
\description{
Title
}
\examples{
kwtest(otutab,metadata[,"Group",drop=F])->res
bbtt(res,pvalue = "p.format")
}
