% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{m_group_env}
\alias{m_group_env}
\alias{plot.mant_g}
\title{Multi-table test with env}
\usage{
m_group_env(g_otutab, env)

\method{plot}{mant_g}(mant_g, env)
}
\arguments{
\item{g_otutab}{multi-otutabs with first column is group}

\item{env}{environmental factors}
}
\value{
a mant_g object

a ggplot
}
\description{
Multi-table test with env

Plot mant_g object
}
\examples{
data(otutab)
cbind(group=rep(c('a','b','c'),c(200,100,192)),otutab)->g_otutab
metadata[,12:16,drop=F]->env
m_group_env(g_otutab,env)->mant_g
plot(mant_g,env)
}
