% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{myRDA}
\alias{myRDA}
\alias{myCCA}
\alias{myCAP}
\title{RDA}
\usage{
myRDA(otutab, env, choose_var = FALSE, norm = TRUE, scale = FALSE, nperm = 499)

myCCA(
  otutab,
  env,
  choose_var = FALSE,
  norm = TRUE,
  scale = FALSE,
  nperm = nperm
)

myCAP(
  otutab,
  env,
  choose_var = FALSE,
  norm = TRUE,
  scale = FALSE,
  dist = "bray"
)
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{env}{environmental factors}

\item{choose_var}{should choose variables? use forward step}

\item{norm}{should normalize?(default:TRUE)}

\item{scale}{should scale species?(default:FALSE)}

\item{nperm}{number of permutation}

\item{dist}{dist method for CAP, see \code{\link[vegan]{vegdist}}}
}
\value{
rda/cca
}
\description{
RDA
}
\examples{
data(otutab, package = "pcutils")
env <- metadata[, 6:10]
# RDA
myRDA(otutab, env) -> phy.rda
RDA_plot(phy.rda, "Group", metadata)
}
\seealso{
\code{\link[vegan]{vegdist};\link[picante]{unifrac}}
}
