% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{name_or_id2df}
\alias{name_or_id2df}
\title{Transfer taxon name or taxid to the lineage dataframe}
\usage{
name_or_id2df(name_or_id, mode = "name")
}
\arguments{
\item{name_or_id}{name or taxid}

\item{mode}{"id" or "name"}
}
\value{
dataframe
}
\description{
Transfer taxon name or taxid to the lineage dataframe
}
\examples{
\dontrun{
name_or_id2df(c("Homo sapiens","Akkermansia muciniphila ATCC BAA-835"))
}
}
