% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_micro.R
\name{ncm}
\alias{ncm}
\title{Sloan Neutral Model}
\usage{
ncm(otutab, model = "nls")
}
\arguments{
\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{model}{fit method, one of "nls","mle"}
}
\value{
ncm_res
}
\description{
Sloan Neutral Model
}
\examples{
data(otutab)
ncm(otutab)->ncm_res
plot(ncm_res)
}
\references{
1. Sloan, W. T. et al. Quantifying the roles of immigration and chance in shaping prokaryote community structure. Environmental Microbiology 8, 732–740 (2006).
}
