% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_tax1.R
\docType{data}
\name{pc_tax1}
\alias{pc_tax1}
\title{testdata (pc_otu class) for pc_tax package.}
\format{
a pc_otu contains an otutable, metadata and a taxomomy table.
\describe{
\item{tbls}{contians otutable rawdata}
\item{metas}{contians metadata}
\item{otus}{contians taxomomy table}
}
}
\usage{
pc_tax1
}
\description{
an otutable, metadata and a taxomomy table.
}
