% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_diversity.R
\name{plot.a_res}
\alias{plot.a_res}
\title{Plot a_res object}
\usage{
\method{plot}{a_res}(x, metadata, group, ...)
}
\arguments{
\item{x}{a a_res object}

\item{metadata}{metadata}

\item{group}{one of colname of metadata}

\item{...}{addditional parameters for \code{\link{group_box}} or \code{\link{my_lm}}}
}
\value{
patchwork object,you can change theme with &
}
\description{
Plot a_res object
}
\seealso{
\code{\link{a_diversity}}
}
