% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot.b_res}
\alias{plot.b_res}
\title{Plot a b_res}
\usage{
\method{plot}{b_res}(
  b_res,
  Group,
  groupname = "level",
  mode = 1,
  bi = F,
  Topn = 10,
  rate = 1,
  margin = F,
  box = T,
  pal = NULL,
  sample_label = T,
  coord_fix = F,
  ...
)
}
\arguments{
\item{b_res}{a b_res object}

\item{Group}{group vector for color}

\item{groupname}{tittle of legend}

\item{mode}{plot mode:1~3}

\item{bi}{plot variables segments?}

\item{Topn}{how many variables to show?}

\item{rate}{segments length rate}

\item{margin}{plot the margin boxplot?}

\item{box}{margin plot box or density?}

\item{pal}{colors for group}

\item{sample_label}{plot the labels of samples?}

\item{coord_fix}{fix the coordinates y/x ratio}

\item{...}{}
}
\value{
a ggplot
}
\description{
Plot a b_res
}
\seealso{
\code{\link{b_analyse}}
}
