% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot.g_test}
\alias{plot.g_test}
\title{Plot g_test}
\usage{
\method{plot}{g_test}(aa)
}
\arguments{
\item{aa}{a g_test object}
}
\value{
ggplot
}
\description{
Plot g_test
}
\seealso{
\code{\link{permanova}}
}
