% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{plot.pro_res}
\alias{plot.pro_res}
\title{Plot pro_res}
\usage{
\method{plot}{pro_res}(pro_res, group, metadata = NULL, pal = NULL)
}
\arguments{
\item{pro_res}{pro_res}

\item{group}{group}

\item{metadata}{metadata}

\item{pal}{pal}
}
\value{
ggplot
}
\description{
Plot pro_res
}
