% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rare_curve_sam}
\alias{rare_curve_sam}
\title{Rare the sample and plot}
\usage{
rare_curve_sam(otutab, rep = 30, count_cutoff = 1)
}
\arguments{
\item{otutab}{otutab}

\item{rep}{repeats number}

\item{count_cutoff}{cutoff to be 0}
}
\description{
Rare the sample and plot
}
\examples{
rare_curve_sam(otutab)
}
