% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_analyse.R
\name{RDA_plot}
\alias{RDA_plot}
\title{Plot RDA res}
\usage{
RDA_plot(
  phy.rda,
  Group,
  metadata = NULL,
  Group2 = NULL,
  env_rate = 1,
  mode = 1,
  tri = FALSE,
  Topn = 10,
  rate = 1,
  margin = FALSE,
  box = TRUE,
  pal = NULL,
  sample_label = TRUE,
  stat_ellipse = TRUE,
  coord_fix = FALSE,
  bi_text_size = 3,
  env_text_param = NULL,
  ...
)
}
\arguments{
\item{phy.rda}{rda/cca object}

\item{Group}{group vector for color}

\item{metadata}{metadata contain Group}

\item{Group2}{mapping point shape}

\item{env_rate}{default 1}

\item{mode}{plot mode:1~3}

\item{tri}{plot variables segments?}

\item{Topn}{how many variables to show?}

\item{rate}{segments length rate}

\item{margin}{plot the margin boxplot?}

\item{box}{margin plot box or density?}

\item{pal}{colors for group}

\item{sample_label}{plot the labels of samples?}

\item{stat_ellipse}{plot the stat_ellipse?}

\item{coord_fix}{fix the coordinates y/x ratio}

\item{bi_text_size}{biplot text size}

\item{env_text_param}{parameters pass to \code{\link[ggplot2]{geom_text}}}

\item{...}{add}
}
\value{
ggplot
}
\description{
Plot RDA res
}
\seealso{
\code{\link{myRDA}}
}
