% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctax_additional.R
\name{sangji_plot}
\alias{sangji_plot}
\title{Plot a sankey}
\usage{
sangji_plot(
  tree,
  top_N = 5,
  notshow = c(),
  intermediate = FALSE,
  width = 3000,
  height = 500,
  ...
)
}
\arguments{
\item{tree}{result from \code{\link{ann_tree}}}

\item{top_N}{each level has top_N}

\item{notshow}{some words you don't want to show}

\item{intermediate}{logical, show the intermediate rank}

\item{width}{width}

\item{height}{height}

\item{...}{look for parameters in \code{\link[sankeyD3]{sankeyNetwork}}}
}
\value{
html widget
}
\description{
Plot a sankey
}
\examples{
\donttest{
if(requireNamespace("sankeyD3") && requireNamespace("tidytree")) {
data(otutab, package = "pcutils")
ann_tree(taxonomy[, c(1, 5, 6, 7)], otutab) -> tree
sangji_plot(tree)
}
}
}
