% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtaxonkit.R
\name{taxonkit_lineage}
\alias{taxonkit_lineage}
\title{Retrieve Taxonomic Lineage using taxonkit}
\usage{
taxonkit_lineage(
  file_path,
  delimiter = ";",
  no_lineage = FALSE,
  show_lineage_ranks = FALSE,
  show_lineage_taxids = FALSE,
  show_name = FALSE,
  show_rank = FALSE,
  show_status_code = FALSE,
  taxid_field = 1,
  text = F
)
}
\arguments{
\item{file_path}{The path to the input file with taxonomic IDs. Or file text (text=TRUE)}

\item{delimiter}{The field delimiter in the lineage (default ";").}

\item{no_lineage}{Logical, indicating whether to exclude lineage information (default: FALSE).}

\item{show_lineage_ranks}{Logical, indicating whether to append ranks of all levels in the lineage (default: FALSE).}

\item{show_lineage_taxids}{Logical, indicating whether to append lineage consisting of taxids (default: FALSE).}

\item{show_name}{Logical, indicating whether to append scientific name (default: FALSE).}

\item{show_rank}{Logical, indicating whether to append rank of taxids (default: FALSE).}

\item{show_status_code}{Logical, indicating whether to show status code before lineage (default: FALSE).}

\item{taxid_field}{The field index of taxid. Input data should be tab-separated (default: 1).}

\item{text}{logical,}
}
\value{
A character vector containing the taxonomic lineage information.
}
\description{
Retrieve Taxonomic Lineage using taxonkit
}
\examples{
\dontrun{
lineage <- taxonkit_lineage("9606\n63221",show_name = TRUE, show_rank = TRUE,text=T)
lineage
}
}
