% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_net.R
\name{zp_plot}
\alias{zp_plot}
\title{Zi-Pi plot of vertexes}
\usage{
zp_plot(go)
}
\arguments{
\item{go}{igraph object after zp_analyse()}
}
\value{
a ggplot object
}
\description{
Zi-Pi plot of vertexes
}
\seealso{
\link{zp_analyse}
}
